pyproject.toml
[tool.poetry]
name = "super-rag"
version = "0.1.0"
description = ""
authors = ["Bastien98 <bastien.moenaert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.4"
pypdf2 = "^3.0.1"
sqlalchemy = "^2.0.36"
rank-bm25 = "^0.2.2"
nltk = "^3.9.1"
asyncio = "^3.4.3"
anthropic = "^0.37.1"
fastapi = "^0.115.4"
instructor = "^1.6.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.0"
numpy = "^1.26.4"
langchain-openai = "^0.2.4"
langchain-ollama = "^0.2.0"
chromadb = "^0.5.16"
langchain-chroma = "^0.1.4"
mysql-connector-python = "^9.1.0"
python-multipart = "^0.0.16"

#pip install llama-parse ( does not work with poetry atm)
sentence-transformers = "^3.2.1"
langgraph = "^0.2.45"
ollama = "^0.3.3"
python-socketio = "^5.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
